mpu6050: rename mpu6050-regs.h to mpu6050_regs.h
